Separar os componentes para uma melhor manutenção e ajuste na função de pesquisa.
